Move generateTestHeaderString to stripe.webhooks
