+++ lib/stripe.js
@@ -28,7 +28,6 @@ var APP_INFO_PROPERTIES = ['name', 'version', 'url', 'partner_id'];
 
 var EventEmitter = require('events').EventEmitter;
 var utils = require('./utils');
-var testUtils = require('../testUtils');
 
 var resourceNamespace = require('./ResourceNamespace');
 
@@ -339,11 +338,6 @@ Stripe.prototype = {
 
   getTelemetryEnabled: function() {
     return this._enableTelemetry;
-  },
-
-  // Generates a header to be used for webhook mocking
-  generateWebhookHeaderString: function(opts) {
-    return testUtils.generateHeaderString.apply(testUtils, opts);
   },
 
   _prepResources: function() {
